+++ src/components/pages/apps/genes/gsea/gsea-plot/gsea-plot-page.tsx
@@ -53,9 +53,10 @@ import { produce } from 'immer'
 
 import { ResizableSidebar } from '@/components/sidebar/resizable-sidebar'
 import { useToolbarTabs } from '@/components/tabs/tab-provider'
-import { useUpdateEffect } from '@/hooks/update-effect'
 import { SVGProvider, useSVG } from '@/providers/svg-provider'
 
+import { useHydratedEffect } from '@/hooks/hydrated-effect'
+import { useUpdateEffect } from '@/hooks/update-effect'
 import { OptsSidebarMenu } from '../../../matcalc/data/opts-sidebar-menu'
 import { UndoShortcuts } from '../../../matcalc/history/undo-shortcuts'
 import {
@@ -78,7 +79,7 @@ export function GseaPlotPage() {
   //const _id = useStableId('gsea-page')
 
   const { settings: edbSettings } = useEdbSettings()
-  const { settings, updateSettings } = useGseaSettings()
+  const { settings, updateSettings, hasHydrated } = useGseaSettings()
   const { setAppInfo } = useAppInfo()
 
   const [search, setSearch] = useState('')
@@ -95,7 +96,7 @@ export function GseaPlotPage() {
 
   const [searchResults, setSearchResults] = useState<IGseaPathway[]>([])
 
-  const { zoom } = useZoom(PLOT_ZOOM_CHANNEL)
+  const { zoom, setZoom } = useZoom(PLOT_ZOOM_CHANNEL)
 
   const [showFileMenu, setShowFileMenu] = useState(false)
   //const [selectAllDatasets, setSelectAllDatasets] = useState(true)
@@ -120,12 +121,20 @@ export function GseaPlotPage() {
   }, [setToolbarTabs])
 
   useUpdateEffect(() => {
-    updateSettings(
-      produce(settings, (draft) => {
-        draft.page.scale = zoom
-      })
-    )
-  }, [zoom])
+    setZoom(zoom)
+  }, [settings.page.scale])
+
+  useHydratedEffect(
+    () => {
+      updateSettings(
+        produce(settings, (draft) => {
+          draft.page.scale = zoom
+        })
+      )
+    },
+    [zoom],
+    hasHydrated
+  )
 
   const searchIndex = useMemo(() => {
     return new Fuse(phenotypes.map((k) => reportsMap[k]!).flat(), {