+++ src/components/pages/apps/genes/gsea/gsea-plot/gsea-settings-store.ts
@@ -185,6 +185,8 @@ export const DEFAULT_GSEA_DISPLAY_PROPS: IGseaDisplayProps = {
 }
 
 export interface IGseaSettingsStore extends IGseaDisplayProps {
+  hasHydrated: boolean
+  setHasHydrated: (hasHydrated: boolean) => void
   updateSettings: (settings: Partial<IGseaDisplayProps>) => void
 }
 
@@ -192,6 +194,13 @@ export const useGseaSettingsStore = create<IGseaSettingsStore>()(
   persist(
     (set) => ({
       ...DEFAULT_GSEA_DISPLAY_PROPS,
+      hasHydrated: false,
+      setHasHydrated: (hasHydrated: boolean) => {
+        set((state) => ({
+          ...state,
+          hasHydrated,
+        }))
+      },
 
       updateSettings: (settings: Partial<IGseaDisplayProps>) => {
         console.log('updating gsea settings', settings)
@@ -204,6 +213,10 @@ export const useGseaSettingsStore = create<IGseaSettingsStore>()(
     {
       name: SETTINGS_KEY, // name in localStorage
       storage: createJSONStorage(() => localStorage),
+      onRehydrateStorage: () => (state) => {
+        console.log('rehydrating gsea settings from storage')
+        state?.setHasHydrated(true)
+      },
     }
   )
 )
@@ -212,14 +225,16 @@ export function useGseaSettings(): {
   settings: IGseaDisplayProps
   updateSettings: (settings: Partial<IGseaDisplayProps>) => void
   reset: () => void
+  hasHydrated: boolean
 } {
   const settings = useGseaSettingsStore((state) => state)
   const updateSettings = useGseaSettingsStore((state) => state.updateSettings)
+  const hasHydrated = useGseaSettingsStore((state) => state.hasHydrated)
 
   function reset() {
     console.log('resetting useGseaSettings to default')
     updateSettings({ ...DEFAULT_GSEA_DISPLAY_PROPS })
   }
 
-  return { settings, updateSettings, reset }
+  return { settings, updateSettings, reset, hasHydrated }
 }