+++ src/components/pages/apps/venn/venn-page.tsx
@@ -63,7 +63,6 @@ import { ToolbarIconButton } from '@/toolbar/toolbar-icon-button'
 import { MonitorDown } from 'lucide-react'
 import { useHistory } from '../matcalc/history/history-provider/history-provider'
 
-import { useUpdateEffect } from '@/hooks/update-effect'
 import { SVGProvider, useSVG } from '@/providers/svg-provider'
 import { useCurrentSheets } from '../matcalc/history/history-provider/history-contexts'
 import APP_INFO from './manifest.json'
@@ -392,7 +391,7 @@ function VennPage() {
     openFile(`Venn Sets`, { sheets: [df] })
   }, [vennElemMap])
 
-  useUpdateEffect(() => {
+  useEffect(() => {
     updateSettings({ scale: zoom })
   }, [zoom])